Add WindowsPerformanceCounter configuration reloads on intervals
Before this change PerformanceCounter instances that would get
dynamically added to a PerformanceCounterCategory after startup
would not be picked up and collected.

This change adds an option to reload the WindowsPerformanceCounter
configuration and thus re-evaluate any new instances that might
have been added since startup or last configuration reload.

Combined with #14 this fixes #12.